* Consolidated the analysis classes in characterization.
