+++ qiskit_experiments/library/calibration/__init__.py
@@ -52,17 +52,6 @@ module.
     RoughXSXAmplitudeCal
     EFRoughXSXAmplitudeCal
 
-Calibration analysis
-====================
-.. autosummary::
-    :toctree: ../stubs/
-    :template: autosummary/analysis.rst
-
-    DragCalAnalysis
-    FineDragAnalysis
-    FineAmplitudeAnalysis
-    RamseyXYAnalysis
-
 Calibrations management
 =======================
 
@@ -76,7 +65,3 @@ from .fine_amplitude import FineAmplitudeCal, FineXAmplitudeCal, FineSXAmplitude
 from .fine_drag_cal import FineDragCal, FineXDragCal, FineSXDragCal
 from .frequency_cal import FrequencyCal
 
-from .analysis.drag_analysis import DragCalAnalysis
-from .analysis.fine_drag_analysis import FineDragAnalysis
-from .analysis.fine_amplitude_analysis import FineAmplitudeAnalysis
-from .analysis.remsey_xy_analysis import RamseyXYAnalysis
